pyproject.toml
[tool.poetry]
name = "omsk-nii-test"
version = "0.1.0"
description = ""
authors = ["Alex Pavlov <yukiu217@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["classic"], version = "^0.31.1"}
python-dotenv = "^1.0.1"
alembic = "^1.13.3"
loguru = "^0.7.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.35"}
asyncpg = "^0.29.0"
pydantic-settings = "^2.5.2"
pytest = "^8.3.3"
sqlalchemy-utils = "^0.41.2"
sqlalchemy-imageattach = "^1.1.0"
python-multipart = "^0.0.12"
bcrypt = "^4.2.0"
phonenumbers = "^8.13.47"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
